Remove mini Holding Analysis graph and trending stocks from Dashboard
